HIPPA Caveat
This product while hitting a few key components of HIPPA compliance, lacks the majority of the necessary features to be considered HIPPA compliant. If this were to be used in a HIPPA setting we could get nailed for at least $1000 per the numerous shortcoming and at a maximum $50,000. Plus we'd have to fix it within 30 days. Not possible.